requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
boto3==1.34.69
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.11
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.9.0
multidict==6.1.0
openai==1.66.3
pip==20.2.3
propcache==0.3.0
pydantic==2.10.6
pydantic-core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
setuptools==49.2.1
sniffio==1.3.1
starlette==0.46.1
tqdm==4.67.1
typing-extensions==4.12.2
urllib3>=1.25.4,<1.27
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0.1
yarl==1.18.3
pytest==7.3.1
pytest-asyncio==0.21.0
